Move primary traits
