+++ helixflow-core/src/lib.rs
@@ -47,3 +47,87 @@ pub enum HelixFlowError {
 }
 
 pub type HelixFlowResult<T> = std::result::Result<T, HelixFlowError>;
+
+pub trait CRUD
+where
+    Self: Sized,
+{
+    fn create<B: Store<Self>>(&self, backend: &B) -> HelixFlowResult<()>;
+    fn get<B: Store<Self>>(backend: &B, id: &Uuid) -> HelixFlowResult<Self>;
+}
+
+/// Methods to store and retrieve `ITEM` in a backend
+pub trait Store<ITEM> {
+    /// Create a new `ITEM` in the backend.
+    ///
+    /// The returned `ITEM` should be the actual stored record from the backend - to allow
+    /// validation by `CRUD<ITEM>::create()`
+    fn create(&self, item: &ITEM) -> HelixFlowResult<ITEM>;
+
+    /// Get an `ITEM` from the backend
+    fn get(&self, id: &Uuid) -> HelixFlowResult<ITEM>;
+}
+
+impl<ITEM> CRUD for ITEM
+where
+    ITEM: HelixFlowItem + PartialEq + Clone,
+{
+    /// Create this item in a given storage backend.
+    fn create<B: Store<ITEM>>(&self, backend: &B) -> HelixFlowResult<()> {
+        let created_item = backend.create(self)?;
+        if &created_item == self {
+            Ok(())
+        } else {
+            Err(HelixFlowError::Mismatch {
+                expected: Box::new(self.clone()),
+                actual: Box::new(created_item),
+            })
+        }
+    }
+
+    /// Get item from `backend` by `id`
+    fn get<B: Store<ITEM>>(backend: &B, id: &Uuid) -> HelixFlowResult<ITEM> {
+        backend.get(id)
+    }
+}
+
+/// A valid usage of a relationship struct, defining acceptable types for left & right.
+///
+/// E.g. to allow `Contains`to be used for `TaskList -> Contains -> Task`:
+/// ```ignore
+/// impl Relationship for Contains<TaskList, Task> {
+///    type Left = TaskList;
+///    type Right = Task;
+/// }
+/// ```
+// TODO: Add derive macro to generate Relationship, Try & FromResidual for valid type pairings
+pub trait Relationship
+where
+    Self: Sized,
+{
+    type Left;
+    type Right;
+}
+
+/// `impl Link<REL> for LEFT` gives `Left Rel:(-> link_type -> Right)`
+pub trait Link
+where
+    Self: Relationship,
+{
+    fn create_linked_item<B: Relate<Self>>(self, backend: &B) -> HelixFlowResult<()>;
+}
+
+pub trait Linkable<REL: Link> {
+    fn link(&self, right: &REL::Right) -> REL;
+    fn get_linked_items<B: Relate<REL>>(
+        &self,
+        backend: &B,
+    ) -> HelixFlowResult<impl Iterator<Item = REL>>;
+}
+
+/// Methods to relate items in a backend
+pub trait Relate<REL: Link> {
+    /// Create and link the related item
+    fn create_linked_item(&self, link: &REL) -> HelixFlowResult<REL>;
+    fn get_linked_items(&self, left: &REL::Left) -> HelixFlowResult<impl Iterator<Item = REL>>;
+}
