+++ helixflow-core/src/task.rs
@@ -10,7 +10,9 @@ use anyhow::anyhow;
 use serde::{Deserialize, Serialize};
 use uuid::{Uuid, uuid};
 
-use crate::{HelixFlowError, HelixFlowItem, HelixFlowResult};
+use crate::{
+    HelixFlowError, HelixFlowItem, HelixFlowResult, Link, Linkable, Relate, Relationship, Store,
+};
 
 impl HelixFlowItem for Task {
     fn as_any(&self) -> &dyn Any {
@@ -73,24 +75,6 @@ impl TaskList {
     }
 }
 
-/// A valid usage of a relationship struct, defining acceptable types for left & right.
-///
-/// E.g. to allow `Contains`to be used for `TaskList -> Contains -> Task`:
-/// ```ignore
-/// impl Relationship for Contains<TaskList, Task> {
-///    type Left = TaskList;
-///    type Right = Task;
-/// }
-/// ```
-// TODO: Add derive macro to generate Relationship, Try & FromResidual for valid type pairings
-pub trait Relationship
-where
-    Self: Sized,
-{
-    type Left;
-    type Right;
-}
-
 impl Relationship for Contains<TaskList, Task> {
     type Left = TaskList;
     type Right = Task;
@@ -149,72 +133,6 @@ where
             },
         })
     }
-}
-
-pub trait CRUD
-where
-    Self: Sized,
-{
-    fn create<B: Store<Self>>(&self, backend: &B) -> HelixFlowResult<()>;
-    fn get<B: Store<Self>>(backend: &B, id: &Uuid) -> HelixFlowResult<Self>;
-}
-
-/// Methods to store and retrieve `ITEM` in a backend
-pub trait Store<ITEM> {
-    /// Create a new `ITEM` in the backend.
-    ///
-    /// The returned `ITEM` should be the actual stored record from the backend - to allow
-    /// validation by `CRUD<ITEM>::create()`
-    fn create(&self, item: &ITEM) -> HelixFlowResult<ITEM>;
-
-    /// Get an `ITEM` from the backend
-    fn get(&self, id: &Uuid) -> HelixFlowResult<ITEM>;
-}
-
-impl<ITEM> CRUD for ITEM
-where
-    ITEM: HelixFlowItem + PartialEq + Clone,
-{
-    /// Create this item in a given storage backend.
-    fn create<B: Store<ITEM>>(&self, backend: &B) -> HelixFlowResult<()> {
-        let created_item = backend.create(self)?;
-        if &created_item == self {
-            Ok(())
-        } else {
-            Err(HelixFlowError::Mismatch {
-                expected: Box::new(self.clone()),
-                actual: Box::new(created_item),
-            })
-        }
-    }
-
-    /// Get item from `backend` by `id`
-    fn get<B: Store<ITEM>>(backend: &B, id: &Uuid) -> HelixFlowResult<ITEM> {
-        backend.get(id)
-    }
-}
-
-/// `impl Link<REL> for LEFT` gives `Left Rel:(-> link_type -> Right)`
-pub trait Link
-where
-    Self: Relationship,
-{
-    fn create_linked_item<B: Relate<Self>>(self, backend: &B) -> HelixFlowResult<()>;
-}
-
-pub trait Linkable<REL: Link> {
-    fn link(&self, right: &REL::Right) -> REL;
-    fn get_linked_items<B: Relate<REL>>(
-        &self,
-        backend: &B,
-    ) -> HelixFlowResult<impl Iterator<Item = REL>>;
-}
-
-/// Methods to relate items in a backend
-pub trait Relate<REL: Link> {
-    /// Create and link the related item
-    fn create_linked_item(&self, link: &REL) -> HelixFlowResult<REL>;
-    fn get_linked_items(&self, left: &REL::Left) -> HelixFlowResult<impl Iterator<Item = REL>>;
 }
 
 impl Link for Contains<TaskList, Task> {
@@ -355,6 +273,8 @@ impl Relate<Contains<TaskList, Task>> for TestBackend {
 #[cfg(test)]
 #[coverage(off)]
 mod tests {
+    use crate::CRUD;
+
     use super::*;
     use std::assert_matches::assert_matches;
