+++ helixflow/src/lib.rs
@@ -5,7 +5,7 @@ use std::{path::PathBuf, rc::Rc};
 use log::debug;
 use slint::ComponentHandle;
 
-use helixflow_core::task::{CRUD, TaskList};
+use helixflow_core::{CRUD, task::TaskList};
 use helixflow_slint::{
     HelixFlow,
     task::{create_task, create_task_in_backlog, load_backlog},
