+++ ui/helixflow-slint/src/task.rs
@@ -6,8 +6,8 @@ use slint::{ComponentHandle, VecModel};
 use slint::{Global, ModelRc, SharedString, ToSharedString};
 
 use helixflow_core::{
-    HelixFlowError, HelixFlowResult,
-    task::{CRUD, Contains, Link, Linkable, Relate, Store, Task, TaskList},
+    CRUD, HelixFlowError, HelixFlowResult, Link, Linkable, Relate, Store,
+    task::{Contains, Task, TaskList},
 };
 
 use crate::{Backlog, CurrentTask, HelixFlow, SlintTask, SlintTaskList};
